Add intial Prototype
Adds files that are included with my intial prototype to get a rhythm system working using metasounds and cue points. At this
stage players can jam out to the beat by pressing the "M" key on the keyboard and their color will change. There is a central WorldBeatController that any actor can subscribe to and receive a message when a beat has occured.