Add UI for teleporting between teleport pads
When stepping onto a teleport pad a series of options are displayed for the player to interact with. When choosing an option the player is teleported to that pad and the UI is removed. Also if the player decides not to teleport and steps off the pad the UI is removed.