Add Teleport Pads and Context Sensitive Button System
adds special tiles that the player can teleport to. This allows them to change levels easily if they've fallen and can't return to a previous level.
A context sensitive prompt has been added that allows the player to interact with the teleport pad. It has been designed to handle many kinds of interactions and can be reused
for other interactions in the future.