Add materials and blueprints for dynamic map
A Dynamic map is created as the player explores the levels. To achieve this I've added a scene capture 2d that writes to a texture, and that texture
will be used as the map in UI. Each tile now has a "sibling" that is captured by the scene capture actor. The sibling is transparent until the player has overlapped
with its corresponding tile on the main map. Sibling tiles are far away from the main map so the player will never be able to see them.