Add blocking panels to show only current level in map
Adds a blocking mesh to the sibling tiles used for creating the map.
The panels block the scene capture from seeing levels below the current one.
The scene capture is moved to the level the player is currently on. In order to make this work
the sibling tiles needed to be moved further away from the actual play area so players could not see it.