Add C++ to project to extend InstancedStaticMesh
Adds a new child class of InstancedStaticMesh that overrides the AddInstance method to add a collision comp
whenever a new instance is added. This comp will also hold all the necessary datatypes for storing information about links
between individual instances and which collision comps belong to them. The idea is there will be one instance of each tile type
in the world and all individual tiles will be instanced from those with their own collision box for detecting overlaps.