Add working prototype of flip map
Adds a new actor component that attaches to level seeds that spawns a grid of
collision boxes that maps 1:1 to the ISM's on the FlipMap actor. When a player
overlaps a box on the map  it will flip the corresponding tile on the flip map
to reveal a large portion of the map at once.