Add music files and use Quartz for beat tracking
Adds a few new wav files I composed for testing quartz as a way of implementing
beat tracking. In the process learned more about how latency can affect gameplay.
So implemented a basic latency tracker that can delay visual effects to line up
with the audio. Also can be used to more accurately determine if the player is
on the beat or not. For now latency tracks as player controls the game, but will
eventually be moved to a calibration setting only.