Add TurnManager Actor
TurnManager keeps track of data related to everything that will happen in a turn.
It will allow the player to scrub through the timeline of a turn to carefully plan
their actions.This commit introduces the turn manager and connects it with the
slider UI allowing the player to see what will happen at a given moment.